Receive typelize_from call even if Typelize is not enabled
Otherwise, there can be a problem if a non-typelizer version of the
Rails app boots up, and then (in the same process) we later try to run
Typelize. In this case, the serializer class won't be re-parsed, so the
main body of the `typelize_from` method won't ever have been called,
making the attempted usage of `typelize_from` not have the intended
effect.